first version of using vga 256 colors simulating mode 13h on dos, it converts RGB to vga colors, but they are still displayed as full RGB colors through SDL

+++ src/display.h
@@ -5,6 +5,6 @@
 
 #include <cstdint>
 
-void AddPixelToBuffer(const DisplayData *dd, int x, int y, uint32_t color);
+void AddPixelToBuffer(const DisplayData *dd, int x, int y, uint8_t color);
 
 #endif
