+++ src/shapes.cpp
@@ -5,7 +5,7 @@
 #include <cmath>
 #include <cstdint>
 
-void AddCircle(const DisplayData *dd, float const radius, Vector2i const circlePos, uint32_t const color) {
+void AddCircle(const DisplayData *dd, float const radius, Vector2i const circlePos, uint8_t const color) {
   int x = static_cast<int>(radius);
   int y = 0;
 
@@ -28,7 +28,7 @@ void AddCircle(const DisplayData *dd, float const radius, Vector2i const circleP
   }
 }
 
-void PlotLineLow(const DisplayData *dd, const Vector2i from, const Vector2i to, const uint32_t color) {
+void PlotLineLow(const DisplayData *dd, const Vector2i from, const Vector2i to, const uint8_t color) {
   const int dx = to.x - from.x;
   int dy = to.y - from.y;
 
@@ -53,7 +53,7 @@ void PlotLineLow(const DisplayData *dd, const Vector2i from, const Vector2i to, 
   }
 }
 
-void PlotLineHigh(const DisplayData *dd, const Vector2i from, const Vector2i to, const uint32_t color) {
+void PlotLineHigh(const DisplayData *dd, const Vector2i from, const Vector2i to, const uint8_t color) {
   int dx = to.x - from.x;
   const int dy = to.y - from.y;
 
@@ -78,7 +78,7 @@ void PlotLineHigh(const DisplayData *dd, const Vector2i from, const Vector2i to,
   }
 }
 
-void AddLine(const DisplayData *dd, Vector2i const from, Vector2i const to, uint32_t const color) {
+void AddLine(const DisplayData *dd, Vector2i const from, Vector2i const to, uint8_t const color) {
   if (abs(to.y - from.y) < abs(to.x - from.x)) {
     if (from.x > to.x)
       PlotLineLow(dd, to, from, color);
@@ -102,7 +102,7 @@ Vector2i CalculateLineEndpoint(Vector2i const from, float const length, float co
   return arrowEndPoint;
 }
 
-void AddLineWithArrow(const DisplayData *dd, Vector2i const from, Vector2i const to, float const rot, uint32_t const color) {
+void AddLineWithArrow(const DisplayData *dd, Vector2i const from, Vector2i const to, float const rot, uint8_t const color) {
   AddLine(dd, from, to, color);
 
   float arrowHeadAngle = deg2rad(135);
@@ -113,12 +113,12 @@ void AddLineWithArrow(const DisplayData *dd, Vector2i const from, Vector2i const
   }
 }
 
-void AddLineInDirectionWithArrow(const DisplayData *dd, Vector2i const from, float const length, float const rot, uint32_t const color) {
+void AddLineInDirectionWithArrow(const DisplayData *dd, Vector2i const from, float const length, float const rot, uint8_t const color) {
   const Vector2i lineEndpoint = CalculateLineEndpoint(from, length, rot);
   AddLineWithArrow(dd, from, lineEndpoint, rot, color);
 }
 
-void AddLineInDirection(const DisplayData *dd, Vector2i const from, float const length, float const rot, uint32_t const color) {
+void AddLineInDirection(const DisplayData *dd, Vector2i const from, float const length, float const rot, uint8_t const color) {
   const Vector2i lineEndpoint = CalculateLineEndpoint(from, length, rot);
   AddLine(dd, from, lineEndpoint, color);
 }
